buscando maximos

+++ PlantillaPractica2SI/src/plantillapractica2si/Hiperplano.java
@@ -16,12 +16,12 @@ public class Hiperplano {
     private static float MaxVector=1;
     private static float MinVector=-1;
     
-    public Hiperplano(float MaxPunto, float MinPunto){
+    public Hiperplano(int MaxPunto, int MinPunto){
         
         int i;
         for (i=0;i<Dimension;i++){
             
-            Punto[i]=(float)(Math.random()*(MaxPunto-MinPunto))+MinPunto;
+            Punto[i]=(int)(Math.random()*(MaxPunto-MinPunto))+MinPunto;
             Vector[i]=(float)(Math.random()*(MaxVector-MinVector))+MinVector;    
         }
     }
